Implement MathUtils; Handle image public id in pet controllers.

+++ src/main/java/com/group12/petweb/controller/api/AdminPetApiController.java
@@ -10,7 +10,6 @@ import javax.servlet.ServletException;
 import javax.servlet.http.*;
 
 import com.cloudinary.Cloudinary;
-import com.google.gson.Gson;
 import com.group12.petweb.dao.PetDao;
 import com.group12.petweb.model.Pet;
 
@@ -38,19 +37,22 @@ public class AdminPetApiController extends HttpServlet {
 			pageSize = 10;
 		}
 		final var pets = petDao.findSomeOffset((page - 1) * 10, pageSize);
-		response.setContentType("application/json");
-		response.getWriter().print(new Gson().toJson(pets));
+		for (final var pet : pets) {
+			pet.setImagePublicId(cloudinary.url().secure(true).publicId(pet.getImagePublicId()).generate());
+		}
 		request.getSession(false).setAttribute("admin.pets", pets);
+		final var props = new HashMap<String, Object>();
+		props.put("pets", pets);
+		props.put("page", page);
+		props.put("pageSize", pets.length);
+		props.put("total", petDao.countAll());
+		request.setAttribute("props", props);
+		request.getRequestDispatcher("/WEB-INF/templates/admin/pet/TableSlot.jsp").forward(request, response);
 	}
 
 	@Override()
 	public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
 		final var part = request.getPart("imageFile");
-		final var fileName = part.getSubmittedFileName();
-		final var index = fileName.indexOf('.');
-		final var name = fileName.substring(0, index);
-		final var ext = fileName.substring(index);
-		final var temp = File.createTempFile(name + '-', ext, new File("/tmp"));
 		try (final var is = part.getInputStream()) {
 			final var model = new Pet();
 			model.setId(UUID.randomUUID());
@@ -64,16 +66,26 @@ public class AdminPetApiController extends HttpServlet {
 				model.setStock(0);
 			}
 
-			Files.copy(is, temp.toPath(), StandardCopyOption.REPLACE_EXISTING);
-			final var res = cloudinary.uploader().upload(
-					temp,
-					new HashMap<String, String>() {
-						{
-							put("public_id", "pet-" + model.getId().toString());
-						}
-					});
-			temp.delete();
-			model.setImageUrl((String) res.get("secure_url"));
+			final var fileName = part.getSubmittedFileName();
+			if (fileName != null && !fileName.isEmpty()) {
+				final var index = fileName.indexOf('.');
+				final var name = fileName.substring(0, index);
+				final var ext = fileName.substring(index);
+				final var temp = File.createTempFile(name + '-', ext, new File("/tmp"));
+				final var publicId = "pet/" + model.getId().toString();
+				Files.copy(is, temp.toPath(), StandardCopyOption.REPLACE_EXISTING);
+				cloudinary.uploader().upload(
+						temp,
+						new HashMap<String, String>() {
+							{
+								put("public_id", publicId);
+							}
+						});
+				temp.delete();
+				model.setImagePublicId(publicId);
+			} else {
+				model.setImagePublicId("pet/default.jpg");
+			}
 			petDao.update(model);
 			response.setStatus(HttpServletResponse.SC_OK);
 			response.getWriter().print(HttpServletResponse.SC_OK);
